requirements.txt
python-dotenv
numpy
streamlit
streamlit-pills
streamlit-toggle-switch
pandas
gspread
google-auth
google-api-python-client
selenium
webdriver_manager